+++ ntia_conformance_checker/ntia_checker.py
@@ -36,13 +36,18 @@ class NTIAChecker(RuleBasedChecker):
 
     _SPEC: "Spec | None" = _NTIA_SPEC
 
-    def __init__(
+    def __init__(  # pylint: disable=too-many-arguments,too-many-positional-arguments
         self,
         file: str,
         validate: bool = True,
         compliance: str = "ntia",
         sbom_spec: str = "spdx2",
+        target_maturity: int = 0,
     ) -> None:
         super().__init__(
-            file=file, validate=validate, compliance=compliance, sbom_spec=sbom_spec
+            file=file,
+            validate=validate,
+            compliance=compliance,
+            sbom_spec=sbom_spec,
+            target_maturity=target_maturity,
         )